Major change to Category and CategoryData
