--- src/CategoryType.ts
@@ -1,14 +1,0 @@
-export namespace CategoryType {
-    // TODO: add to CategoryData somehow? use in Node.isData() etc?
-    export enum Type {
-        Application = "Application",
-        Container = "Container",
-        Control = "Control",
-        Data = "Data",
-        Construct = "Construct",
-        Other = "Other",
-        Service = "Service",
-        Socket = "Socket",
-        Unknown = "Unknown",
-    }
-}
