+++ tsconfig.json
@@ -22,7 +22,8 @@
         "./src/components.ts",
         "./src/main.ts",
         "./src/Config.ts",
-        "./src/CategoryType.ts",
+        "./src/Category.ts",
+        "./src/CategoryData.ts",
         "./src/Eagle.ts",
         "./src/GitHub.ts",
         "./src/GitLab.ts",
